Simplify Agents catalog and consolidate run history into Monitoring
Agents catalog (/agents) drops its internal tabs and only lists
AgentDefinition-backed entries (no more bare plugin cards), titled to
match Models. Run History moves into Monitoring -> Agents, replacing
the old per-agent/recent-runs/errors summaries with the same
AgentRunListTable + filters used previously, so nothing is lost.

Also corrects CONTEXT.md's Agent glossary entry: foundationModel and
skills were documented as bundled but neither is wired at runtime
(model selection is step-level; there's no skills field on the
schema) -- now flagged as aspirational rather than current behaviour.

Includes the seed:dev script fix (npx -> pnpm exec tsx) needed to get
local dev running.
