+++ packages/platform-ui/e2e/ui/agent-oversight.journey.ts
@@ -8,17 +8,13 @@ test.describe('Agent Oversight Journey', () => {
     test.setTimeout(60_000); // multiple navigations with async API loading
     trackPageErrors(page);
     await page.goto(`/${TEST_ORG_HANDLE}/agents`);
-    await expect(page.getByText('Available AI agents for building workflows')).toBeVisible({ timeout: 10_000 });
-
-    // "Available Agents" tab is the default — wait for the Run History tab to be
-    // present, which confirms the page shell rendered. The catalog content depends
-    // on plugin infrastructure that may not be present in emulator mode, so we
-    // don't assert on specific catalog items here.
-    await expect(page.getByRole('tab', { name: 'Run History' })).toBeVisible({ timeout: 10_000 });
+    await expect(page.getByText('Custom Agents')).toBeVisible({ timeout: 10_000 });
+    await expect(page.getByText('Custom configuration for foundation models')).toBeVisible();
     await expect(page.getByRole('link', { name: 'New Agent', exact: true })).toBeVisible();
 
-    // Switch to Run History tab
-    await page.getByRole('tab', { name: 'Run History' }).click();
+    // Run history lives on Monitoring → Agents, not on the Agents catalog page.
+    await page.goto(`/${TEST_ORG_HANDLE}/monitoring`);
+    await page.getByRole('tab', { name: 'Agents' }).click();
     await expect(page.getByText('Narrative Summary').first()).toBeVisible({ timeout: 10_000 });
 
     // Autonomy column shows control mode labels, not raw L-level codes
@@ -56,7 +52,7 @@ test.describe('Agent Oversight Journey', () => {
   test('create a new agent and verify redirect', async ({ page }) => {
     trackPageErrors(page);
     await page.goto(`/${TEST_ORG_HANDLE}/agents`);
-    await expect(page.getByText('Available AI agents for building workflows')).toBeVisible({ timeout: 10_000 });
+    await expect(page.getByText('Custom Agents')).toBeVisible({ timeout: 10_000 });
     await page.getByRole('link', { name: 'New Agent', exact: true }).click();
     await page.waitForURL(`**/${TEST_ORG_HANDLE}/agents/new`, { timeout: 20_000 });
     await expect(page.getByText('Register a new AI agent and configure its capabilities.')).toBeVisible({ timeout: 10_000 });
