+++ packages/platform-ui/package.json
@@ -15,7 +15,7 @@
     "test:e2e": "NODE_OPTIONS='--conditions=@mediforce/source' E2E_FULL_SUITE=true playwright test",
     "test:e2e:headed": "NODE_OPTIONS='--conditions=@mediforce/source' E2E_FULL_SUITE=true E2E_DEV_SERVER=true playwright test --headed",
     "test:e2e:ui": "NODE_OPTIONS='--conditions=@mediforce/source' E2E_DEV_SERVER=true playwright test --ui",
-    "seed:dev": "npx tsx scripts/seed-dev-data.ts"
+    "seed:dev": "pnpm exec tsx scripts/seed-dev-data.ts"
   },
   "dependencies": {
     "@auth/drizzle-adapter": "1.11.2",
